Testing structures

+++ my-spring-boot/src/main/java/org/netflix/data_structures/Node.java
@@ -1,3 +1,3 @@
 package org.netflix.data_structures;
 
-public record Node(String label, int weight) {}
+public record Node(String label, long weight) {}
